requirements.txt
discord.py
python-dotenv
pymongo
dnspython
certifi
